move start debugger to DebugHostController
